+++ src/main/java/com/teamgeso/hotelback/controller/RoomTypeController.java
@@ -6,7 +6,8 @@ import com.teamgeso.hotelback.model.RoomType;
 import com.teamgeso.hotelback.repository.RoomTypeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-
+import org.springframework.http.ResponseEntity;
+import org.springframework.http.HttpStatus;
 
 import java.util.Optional;
 
@@ -32,30 +33,54 @@ public class RoomTypeController implements DaoRoomType {
 
     @PostMapping(value = "")
     @ResponseBody
-    public Optional<RoomType> createRoomType(@RequestBody RoomTypeDTO roomType){
-        RoomType roomTypeToSave = new RoomType();
-        roomTypeToSave.setName(roomType.getName());
-        roomTypeToSave.setPrice(roomType.getPrice());
-        roomTypeRepository.save(roomTypeToSave);
-        return roomTypeRepository.findById(roomTypeToSave.getId());
+    public ResponseEntity createRoomType(@RequestBody RoomTypeDTO roomType){
+        RoomType createdRoomType = new RoomType();
+        createdRoomType.setName(roomType.getName());
+        createdRoomType.setPrice(roomType.getPrice());
+        createdRoomType.setCapacity(roomType.getCapacity());
+
+        if (createdRoomType.getName() != null && createdRoomType.getPrice() != null && createdRoomType.getCapacity() != null){
+            if (createdRoomType.getPrice() <= 0 || createdRoomType.getCapacity() <= 0)
+                return new ResponseEntity<>("El precio y/o la capacidad de la habitación son menores o iguales a 0", HttpStatus.BAD_REQUEST);
+
+            return new ResponseEntity<>(roomTypeRepository.save(createdRoomType),HttpStatus.CREATED);
+        }
+
+        return new ResponseEntity<>("El tipo de habitación no puede contener valores nulos.", HttpStatus.BAD_REQUEST);
     }
 
-    @PostMapping (value = "/{id}")
+    @PutMapping (value = "/{id}")
     public @ResponseBody
-    Optional<RoomType> updateRoomType(@PathVariable Integer id, @RequestBody RoomTypeDTO roomType){
-        RoomType updatedRoomType = new RoomType();
-        updatedRoomType.setName(roomType.getName());
-        updatedRoomType.setPrice(roomType.getPrice());
-        updatedRoomType.setId(id);
-        roomTypeRepository.save(updatedRoomType);
-        return roomTypeRepository.findById(updatedRoomType.getId());
+    ResponseEntity updateRoomType(@PathVariable Integer id, @RequestBody RoomTypeDTO roomType){
+        RoomType roomTypeToUpdate = roomTypeRepository.findRoomTypeById(id);
+        if (roomTypeToUpdate == null)
+            return new ResponseEntity<>("La habitación a editar no se ha podido encontrar.", HttpStatus.BAD_REQUEST);
+
+        roomTypeToUpdate.setPrice(roomType.getPrice());
+        roomTypeToUpdate.setName(roomType.getName());
+        roomTypeToUpdate.setCapacity(roomType.getCapacity());
+
+        if (roomTypeToUpdate.getPrice() != null && roomTypeToUpdate.getName() != null && roomTypeToUpdate.getCapacity() != null){
+            if (roomTypeToUpdate.getPrice() < 0 || roomTypeToUpdate.getCapacity() <= 0)
+                return new ResponseEntity<>("El precio y/o la capacidad de la habitación son menores o iguales a 0", HttpStatus.BAD_REQUEST);
+
+            return new ResponseEntity<>(roomTypeRepository.save(roomTypeToUpdate),HttpStatus.CREATED);
+        }
+
+        return new ResponseEntity<>("Un valor no puede ser modificado por un valor nulo.", HttpStatus.BAD_REQUEST);
     }
 
     @PostMapping (value = "/{id}/delete")
     public @ResponseBody
-    Iterable<RoomType> deleteRoomType(@PathVariable Integer id){
-        roomTypeRepository.deleteById(id);
-        return roomTypeRepository.findAll();
+    ResponseEntity deleteRoomType(@PathVariable Integer id){
+        RoomType roomTypeToDelete = roomTypeRepository.findRoomTypeById(id);
+        
+        if (roomTypeToDelete != null){
+            roomTypeRepository.deleteById(id);
+            return new ResponseEntity<>("Borrado exitosamente",HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>("El tipo de habitación a borrar no existe.", HttpStatus.BAD_REQUEST);
     }
